requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
openai>=1.86.0
qdrant-client>=1.16.1
langchain-qdrant>=0.2.1
tenacity>=8.5.0
requests==2.32.3
unstructured[md]==0.16.1
tiktoken==0.8.0
nltk==3.9.1
httpx==0.27.0
python-dateutil==2.9.0.post0
langfuse>=2.53.0
chainlit>=1.3.0
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
fastapi==0.115.14
uvicorn==0.39.0
redis>=5.0.0
